requirements.txt
Django==4.2.27
djangorestframework==3.16.1
gunicorn==23.0.0
requests
google-generativeai
